Fix TeamNotFoundException
