+++ src/main/java/seedu/address/logic/Messages.java
@@ -6,7 +6,6 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.parser.Prefix;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.Team;
 
 /**
  * Container for user visible messages.
@@ -50,13 +49,4 @@ public class Messages {
         person.getTags().forEach(builder::append);
         return builder.toString();
     }
-
-    /**
-     * Formats the {@code team} for display to the user.
-     */
-    public static String format(String teamName) {
-        //Implement this
-        return teamName;
-    }
-
 }
