+++ src/main/java/seedu/address/logic/parser/DeleteTeamParser.java
@@ -1,27 +1,47 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_TEAMNAME;
 
-import seedu.address.logic.commands.DeleteTeam;
+import java.util.stream.Stream;
+
+import seedu.address.logic.commands.DeleteTeamCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+
 /**
- * Parses input arguments and creates a new DeleteCommand object
+ * Parses input arguments and creates a new DeleteTeamCommand object
  */
-public class DeleteTeamParser implements Parser<DeleteTeam> {
+public class DeleteTeamParser implements Parser<DeleteTeamCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the DeleteCommand
      * and returns a DeleteCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public DeleteTeam parse(String args) throws ParseException {
+    public DeleteTeamCommand parse(String args) throws ParseException {
+        ArgumentMultimap argMultimap =
+                ArgumentTokenizer.tokenize(args, PREFIX_TEAMNAME);
+
+        if (!arePrefixesPresent(argMultimap, PREFIX_TEAMNAME)
+                || !argMultimap.getPreamble().isEmpty()) {
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteTeamCommand.MESSAGE_USAGE));
+        }
+        argMultimap.verifyNoDuplicatePrefixesFor(PREFIX_TEAMNAME);
         try {
-            String teamName = ParserUtil.parseTeamName(args);
-            return new DeleteTeam(teamName);
+            String teamName = ParserUtil.parseTeamName(argMultimap.getValue(PREFIX_TEAMNAME).get());
+            return new DeleteTeamCommand(teamName);
         } catch (ParseException pe) {
             throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteTeam.MESSAGE_USAGE), pe);
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteTeamCommand.MESSAGE_USAGE), pe);
         }
+    }
+
+    /**
+     * Returns true if none of the prefixes contains empty {@code Optional} values in the given
+     * {@code ArgumentMultimap}.
+     */
+    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix prefix) {
+        return Stream.of(prefix).anyMatch(prefixes -> argumentMultimap.getValue(prefix).isPresent());
     }
 }
