+++ src/main/java/seedu/address/model/AddressBook.java
@@ -7,8 +7,8 @@ import java.util.List;
 import javafx.collections.ObservableList;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.Team;
 import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.team.Team;
 
 /**
  * Wraps all data at the address-book level
