+++ src/main/java/seedu/address/model/ReadOnlyAddressBook.java
@@ -2,7 +2,7 @@ package seedu.address.model;
 
 import javafx.collections.ObservableList;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.Team;
+import seedu.address.model.team.Team;
 
 /**
  * Unmodifiable view of an address book
