+++ src/main/java/seedu/address/model/team/exceptions/TeamNotFoundException.java
@@ -1,0 +1,10 @@
+package seedu.address.model.team.exceptions;
+
+/**
+ * Signals that the operation is unable to find the specified team.
+ */
+public class TeamNotFoundException extends RuntimeException {
+    public TeamNotFoundException() {
+        super("The specified team could not be found");
+    }
+}
