+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -15,9 +15,8 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
 import seedu.address.model.person.Remark;
-import seedu.address.model.person.Team;
 import seedu.address.model.tag.Tag;
-
+import seedu.address.model.team.Team;
 
 
 /**
